bugfix/EPMRPP-37970 Attachment rotation fix
