+++ app/localization/translated/be.json
@@ -83,12 +83,10 @@
   "ApiPage.title": "Дакументацыя API",
   "AssignedProjectsBlock.headerNameCol": "Прызначаныя праекты",
   "AssignedProjectsBlock.headerRoleCol": "Праектная роля",
-  "AttachmentDialog.close": "Close",
-  "AttachmentDialog.editor": "Editor",
-  "AttachmentDialog.errorFileStructure": "Bad file structure",
-  "AttachmentDialog.noAttachmentsMessage": "No attachments to display",
-  "AttachmentDialog.rotate": "Rotate",
-  "AttachmentDialog.title": "Attachment",
+  "AttachmentModal.errorFileStructure": "Няправільная структура файла",
+  "AttachmentModal.rotate": "Круціць",
+  "AttachmentModal.title": "Укладанне",
+  "Attachments.noAttachmentsMessage": "Няма ўкладанняў для адлюстравання",
   "AttributeEditor.attributeKeyLengthHint": "Ключ атрыбуту павінен мець памер ад 1 да 128",
   "AttributeEditor.attributeValueLengthHint": "Значэнне атрыбуту павінна мець памер ад 1 да 128",
   "AttributeList.addNew": "Add new",
@@ -140,6 +138,7 @@
   "Common.add": "Дадаць",
   "Common.cancel": "Адмяніць",
   "Common.clone": "Кланаваць",
+  "Common.close": "Закрыць",
   "Common.closeModalWarning": "Вы павінны захаваць змены або адмяніць іх перад закрыццём вокны",
   "Common.confirm": "Пацвердзіць",
   "Common.delete": "Выдаліць",
