+++ app/localization/translated/ru.json
@@ -83,12 +83,10 @@
   "ApiPage.title": "Документация API",
   "AssignedProjectsBlock.headerNameCol": "Назначенные проекты",
   "AssignedProjectsBlock.headerRoleCol": "Проектная роль",
-  "AttachmentDialog.close": "Close",
-  "AttachmentDialog.editor": "Editor",
-  "AttachmentDialog.errorFileStructure": "Bad file structure",
-  "AttachmentDialog.noAttachmentsMessage": "No attachments to display",
-  "AttachmentDialog.rotate": "Rotate",
-  "AttachmentDialog.title": "Attachment",
+  "AttachmentModal.errorFileStructure": "Неверная структура файла",
+  "AttachmentModal.rotate": "Вращать",
+  "AttachmentModal.title": "Вложение",
+  "Attachments.noAttachmentsMessage": "Нет вложений для отображения",
   "AttributeEditor.attributeKeyLengthHint": "Ключ атрибута должен иметь размер от 1 до 128",
   "AttributeEditor.attributeValueLengthHint": "Значение атрибута должно иметь размер от 1 до 128",
   "AttributeList.addNew": "Add new",
@@ -140,6 +138,7 @@
   "Common.add": "Добавить",
   "Common.cancel": "Отменить",
   "Common.clone": "Клонировать",
+  "Common.close": "Закрыть",
   "Common.closeModalWarning": "Вы должны сохранить изменения или отменить их перед закрытием окна",
   "Common.confirm": "Подтвердить",
   "Common.delete": "Удалить",
