+++ app/src/pages/inside/logsPage/logItemInfo/logItemInfoTabs/attachments/index.js
@@ -1,2 +1,1 @@
 export { Attachments } from './attachments';
-export { AttachmentCodeModal, AttachmentHarFileModal, AttachmentImageModal } from './modals';
