--- app/src/pages/inside/logsPage/logItemInfo/logItemInfoTabs/attachments/modals/messages.js
@@ -1,28 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export const messages = defineMessages({
-  title: {
-    id: 'AttachmentDialog.title',
-    defaultMessage: 'Attachment',
-  },
-  close: {
-    id: 'AttachmentDialog.close',
-    defaultMessage: 'Close',
-  },
-  descriptionPlaceholder: {
-    id: 'AttachmentDialog.editor',
-    defaultMessage: 'Editor',
-  },
-  rotate: {
-    id: 'AttachmentDialog.rotate',
-    defaultMessage: 'Rotate',
-  },
-  noAttachmentsMessage: {
-    id: 'AttachmentDialog.noAttachmentsMessage',
-    defaultMessage: 'No attachments to display',
-  },
-  errorFileStructure: {
-    id: 'AttachmentDialog.errorFileStructure',
-    defaultMessage: 'Bad file structure',
-  },
-});
