+++ app/src/pages/inside/logsPage/modals/attachments/attachmentImageModal.scss
