+++ app/src/pages/inside/logsPage/modals/attachments/index.js
