+++ app/src/pages/inside/logsPage/modals/attachments/messages.js
@@ -1,0 +1,16 @@
+import { defineMessages } from 'react-intl';
+
+export const messages = defineMessages({
+  title: {
+    id: 'AttachmentModal.title',
+    defaultMessage: 'Attachment',
+  },
+  rotate: {
+    id: 'AttachmentModal.rotate',
+    defaultMessage: 'Rotate',
+  },
+  errorFileStructure: {
+    id: 'AttachmentModal.errorFileStructure',
+    defaultMessage: 'Bad file structure',
+  },
+});
