+++ app/src/pages/inside/logsPage/modals/index.js
@@ -1,0 +1,2 @@
+export { CopySendDefectModal } from './copySendDefectModal';
+export { AttachmentImageModal, AttachmentCodeModal, AttachmentHarFileModal } from './attachments';
